# Conflicts:
#	src/main/resources/com/example/legend_of_sparta/spartan-view.fxml

added getter and setter in Player.java (alana)

+++ src/main/java/com/example/legend_of_sparta/Entity.java
@@ -2,5 +2,5 @@ package com.example.legend_of_sparta;
 
 public class Entity
 {
-    int xPos, yPos;
+    double xPos, yPos;
 }
